+++ R/datasets.R
@@ -1,0 +1,18 @@
+#' Weight gain data.
+#'
+#' Data from chapter 11 in Mrode 2005. The variables are as follows:
+#'
+#' \itemize{
+#'   \item calf. a factor with levels \code{4} \code{5} \code{6} \code{7} \code{8}
+#'   \item dam. a factor with levels \code{2} \code{5} \code{6}
+#'   \item sire. a factor with levels \code{1} \code{3} \code{4}
+#'   \item sex. a factor with levels \code{male} \code{female}
+#'   \item WWG11. a numeric vector
+#' }
+#'
+#' @source Mrode. 2005.
+#' @keywords datasets
+#' @examples
+#' data(Mrode11)
+"Mrode11"
+