+++ NEWS.md
@@ -1,7 +1,17 @@
-# X.XX.X
+# 0.1.0.0
 ## NEW
+  - methods for `gremlin` objects
+    - notably, `AIC`, `residuals`, `anova`, and `nobs`
+    - updated the `summary`, `print`, and `logLik` methods as well
 
-## Small changes
+# 0.0.2.0
+Improved algorithm that reduces computational resources and time! Also implemented c++ code in `gremlin()`, while keeping `gremlinR()` purely the R implementation (at least from the package writing standpoint).
+
+# 0.0.1.0
+## NEW
+Documentation has switched from filling out the `.Rd` files manually to providing
+documentation next to the function code in the `.R` files using `roxygen2`
+
 
 # 0.0.0.1 April 2017 `gremlin` is born!